[riscv64] Fix IsIllegalInstruction function
Check whether the right most 16 bits are all-zero.

Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/2983839
Commit-Queue: Ji Qiu <qiuji@iscas.ac.cn>
Cr-Commit-Position: refs/heads/master@{#75343}

+++ src/codegen/riscv64/constants-riscv64.h
@@ -738,7 +738,7 @@ class InstructionBase {
   };
 
   inline bool IsIllegalInstruction() const {
-    uint8_t FirstHalfWord = *reinterpret_cast<const uint16_t*>(this);
+    uint16_t FirstHalfWord = *reinterpret_cast<const uint16_t*>(this);
     return FirstHalfWord == 0;
   }
